akademickisklepzgadzetami/akademickisklepzgadzetami/basket/models.py
from django.db import models
from oscar.apps.basket.abstract_models import AbstractBasket, AbstractLine

class Basket(AbstractBasket):
    pass

class Line(AbstractLine):
    pass

from oscar.apps.basket.models import *