requirements.txt
asgiref==3.10.0
babel==2.17.0
certifi==2025.10.5
charset-normalizer==3.4.4
Django==5.2.7
django-browser-reload==1.21.0
django-extra-views==0.15.0
django-haystack==3.3.0
django-oscar==4.0
django-phonenumber-field==8.3.0
django-tables2==2.7.0
django-tailwind==4.4.2
django-treebeard==4.7.1
django-widget-tweaks==1.5.0
factory_boy==3.3.3
Faker==37.11.0
idna==3.11
packaging==25.0
phonenumbers==9.0.16
pillow==12.0.0
purl==1.6
pysolr==3.10.0
pytailwindcss==0.3.0
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sorl-thumbnail==12.11.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
pycountry==24.6.1
whoosh==2.7.4